requirements.txt
django
asyncio
aiohttp
lxml
requests
datetime
tqdm
django-bootstrap-v5
psycopg2-binary
uvicorn
gunicorn
python-dotenv
whitenoise
mail-parser
glom
django-simple-history
graphene-django
django-filter
strawberry-graphql-django
strawberry-django-auth